Fix inboundAppMessages not emitting early messages
when consumeAsFlow is called, it immediately hogs all the messages, which means that the Buffer in the Channel above has no effect - if nobody is collecting the flow, the messages are lost. This is fixed by moving away from SharedFlow to regular flow and by using receive instead of consume.

A side effect of this is that this flow cannot be collected by more than one consumer at once, but from what I can see, there is no use case for that?